refactor(request): extract pure request domain helpers
